LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.lang.default-expected.txt

2d.text.lang.default
Actual output:

FAIL Canvas test: 2d.text.lang.default assert_equals: ctx.lang === 'inherit' (got [undefined], expected inherit[string]) expected (string) "inherit" but got (undefined) undefined
